get up to date: localization, intra exp merge, dynamic detection

+++ main/src/vtr_lidar/include/vtr_lidar/modules/pointmap/dynamic_detection_module.hpp
@@ -45,7 +45,7 @@ class DynamicDetectionModule : public tactic::BaseModule {
     // dependencies
     std::string intra_exp_merging = IntraExpMergingModule::static_name;
 
-    int depth = 0;
+    double depth = 0;
 
     float horizontal_resolution = 0.001;
     float vertical_resolution = 0.001;
